Refactor enterpriseBotTestBase name to botTestBase
