+++ templates/Enterprise-Template/src/typescript/enterprise-bot/test/flow/escalateDialogTest.js
@@ -1,14 +1,14 @@
-const enterpriseBotTestBase = require('./enterpriseBotTestBase.js');
+const botTestBase = require('./botTestBase.js');
 const testNock = require('../testBase');
 
 describe("Escalate Dialog", function () {
     beforeEach(function () {
-        enterpriseBotTestBase.initialize();
+        botTestBase.initialize();
     });
 
     describe("Escalating", function () {
         it("Send 'I want to talk to a human' and check you get the expected response", function (done) {
-            const testAdapter = enterpriseBotTestBase.getTestAdapter();
+            const testAdapter = botTestBase.getTestAdapter();
             const flow = testAdapter
                 .send('I want to talk to a human')
                 .assertReply('Our agents are available 24/7 at 1(800)555-1234.');
