+++ templates/Enterprise-Template/src/typescript/enterprise-bot/test/flow/mainDialogTest.js
@@ -1,16 +1,16 @@
 const assert = require('assert');
-const enterpriseBotTestBase = require('./enterpriseBotTestBase.js');
+const botTestBase = require('./botTestBase.js');
 const testNock = require('../testBase');
 const introJson = require('../../src/dialogs/main/resources/Intro.json');
 
 describe("Main Dialog", function () {
 	beforeEach(function () {
-		enterpriseBotTestBase.initialize();
+		botTestBase.initialize();
 	});
 
 	describe("Intro Card", function () {
 		it("Send conversationUpdate and verify card is received", function (done) {
-			const testAdapter = enterpriseBotTestBase.getTestAdapter();
+			const testAdapter = botTestBase.getTestAdapter();
 			const flow = testAdapter
 				.send({
 					type: "conversationUpdate",
@@ -36,7 +36,7 @@ describe("Main Dialog", function () {
 
 	describe("Greeting", function () {
 		it("Send Hello and check you get the expected response", function (done) {
-			const testAdapter = enterpriseBotTestBase.getTestAdapter();
+			const testAdapter = botTestBase.getTestAdapter();
 			const flow = testAdapter
 				.send('Hi')
 				.assertReply('Hello.');
@@ -47,7 +47,7 @@ describe("Main Dialog", function () {
 
 	describe("Help", function () {
 		it("Send Help and check you get the expected response", function (done) {
-			const testAdapter = enterpriseBotTestBase.getTestAdapter();
+			const testAdapter = botTestBase.getTestAdapter();
 			const flow = testAdapter
 				.send('Help')
 				.assertReply('This card can be used to display information to help your user interact with your bot. The buttons below can be used for sample queries or links to external sites.');
@@ -58,7 +58,7 @@ describe("Main Dialog", function () {
 
 	describe("Localization", function () {
 		it("Send a message in FIGS+ZH, set locale property on activity and validate the localized response", function (done) {
-			const testAdapter = enterpriseBotTestBase.getTestAdapter();
+			const testAdapter = botTestBase.getTestAdapter();
 			const flow = testAdapter
 				.send({
 					type: "message",
